+++ src/sreg/models/__init__.py
@@ -2,8 +2,13 @@
 
 from sreg.models.case_plan import CasePlan, EvalQuestionPlan
 from sreg.models.dag_spec import DAGNodeSpec, DAGSpec
-from sreg.models.episode import Action, ActionType, Episode, Observation, StepResult
-from sreg.models.research_problem import AvailableAction, DataAsset, ResearchProblem
+from sreg.models.episode import Action, ActionDef, ActionType, Episode, Observation, StepResult
+from sreg.models.research_problem import (
+    AvailableAction,
+    DataAsset,
+    ResearchActionType,
+    ResearchProblem,
+)
 from sreg.models.score import Score, StepScore
 from sreg.models.task import Task, TaskBundle, TaskSpec, TaskType
 from sreg.models.teacher import TeacherOutput
@@ -11,6 +16,7 @@ from sreg.models.world import CPD, DifficultyProfile, Edge, Node, NodeType, Worl
 
 __all__ = [
     "Action",
+    "ActionDef",
     "ActionType",
     "AvailableAction",
     "CPD",
@@ -25,6 +31,7 @@ __all__ = [
     "Node",
     "NodeType",
     "Observation",
+    "ResearchActionType",
     "ResearchProblem",
     "Score",
     "StepResult",